Exercise 4.21

+++ Osa4/controllers/blogs.js
@@ -48,9 +48,22 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== user._id.toString())
+      return response.status(401).json({ error: 'not authorized' })
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
+    if (exception.name === 'JsonWebTokenError')
+      return response.status(401).json({ error: 'token missing or invalid' })
     response.status(400).json({ error: 'malformatted id' })
   }
 })
